Fix degrees to DMH conversion for APRS to support longitudes > 127 and < -128. Use 16-bit signed values for degrees.

+++ src/codecs/aprs/aprs.c
@@ -4,19 +4,19 @@
 
 volatile uint16_t aprs_packet_counter = 0;
 
-void convert_degrees_to_dmh(long x, int8_t *degrees, uint8_t *minutes, uint8_t *h_minutes)
+void convert_degrees_to_dmh(long x, int16_t *degrees, uint8_t *minutes, uint8_t *h_minutes)
 {
     uint8_t sign = (uint8_t) (x > 0 ? 1 : 0);
     if (!sign) {
         x = -(x);
     }
-    *degrees = (int8_t) (x / 1000000);
+    *degrees = (int16_t) (x / 1000000);
     x = x - (*degrees * 1000000);
     x = (x) * 60 / 10000;
     *minutes = (uint8_t) (x / 100);
     *h_minutes = (uint8_t) (x - (*minutes * 100));
     if (!sign) {
-        *degrees = -*degrees;
+        *degrees = (int16_t ) -*degrees;
     }
 }
